stuff for running on -device

+++ release/create_prebuilt.sh
@@ -11,12 +11,24 @@ rm -f panda/board/obj/panda_h7.bin.signed
 VERSION=$(cat common/version.h | awk -F[\"-]  '{print $2}')
 echo "#define COMMA_VERSION \"$VERSION-release\"" > common/version.h
 
+echo "[-] committing version $VERSION T=$SECONDS"
+git add -f .
+git commit -a -m "openpilot v$VERSION release"
+
 # Build
 export PYTHONPATH="$TARGET_DIR"
 scons -j$(nproc)
 
 # release panda fw
 CERT=/data/pandaextra/certs/release RELEASE=1 scons -j$(nproc) panda/
+
+# Ensure no submodules in release
+if test "$(git submodule--helper list | wc -l)" -gt "0"; then
+  echo "submodules found:"
+  git submodule--helper list
+  exit 1
+fi
+git submodule status
 
 # Cleanup
 find . -name '*.a' -delete
